continue lib api cleanup

+++ libfreeipmi/include/freeipmi/util/ipmi-cipher-suite-util.h
@@ -84,15 +84,15 @@ extern "C" {
     || ((__id) >= 6 && (__id) <= 8)          \
     || ((__id) >= 11 && (__id) <= 12)) ? 1 : 0)
 
-int8_t ipmi_cipher_suite_id_to_algorithms (uint8_t cipher_suite_id,
-                                           uint8_t *authentication_algorithm,
-                                           uint8_t *integrity_algorithm,
-                                           uint8_t *confidentiality_algorithm);
+int ipmi_cipher_suite_id_to_algorithms (uint8_t cipher_suite_id,
+                                        uint8_t *authentication_algorithm,
+                                        uint8_t *integrity_algorithm,
+                                        uint8_t *confidentiality_algorithm);
 
-int8_t ipmi_algorithms_to_cipher_suite_id (uint8_t authentication_algorithm,
-                                           uint8_t integrity_algorithm,
-                                           uint8_t confidentiality_algorithm,
-                                           uint8_t *cipher_suite_id);
+int ipmi_algorithms_to_cipher_suite_id (uint8_t authentication_algorithm,
+                                        uint8_t integrity_algorithm,
+                                        uint8_t confidentiality_algorithm,
+                                        uint8_t *cipher_suite_id);
 
 #ifdef __cplusplus
 }
