+++ libfreeipmi/include/freeipmi/util/ipmi-ipmb-util.h
@@ -26,12 +26,12 @@ extern "C" {
 #include <stdint.h>
 #include <freeipmi/fiid/fiid.h>
 
-int8_t ipmi_ipmb_check_rq_seq (fiid_obj_t obj_ipmb_msg_hdr, uint8_t rq_seq);
+int ipmi_ipmb_check_rq_seq (fiid_obj_t obj_ipmb_msg_hdr, uint8_t rq_seq);
 
-int8_t ipmi_ipmb_check_checksum (uint8_t rq_addr,
-                                 fiid_obj_t obj_ipmb_msg_hdr,
-                                 fiid_obj_t obj_cmd,
-                                 fiid_obj_t obj_ipmb_msg_trlr);
+int ipmi_ipmb_check_checksum (uint8_t rq_addr,
+                              fiid_obj_t obj_ipmb_msg_hdr,
+                              fiid_obj_t obj_cmd,
+                              fiid_obj_t obj_ipmb_msg_trlr);
 
 #ifdef __cplusplus
 }
